Move model-specific reasoning parser tests
